HDT connector in connectors class
